requirements.txt
### WS
django
psycopg2-binary
bootstrap4
six
boto3==1.28.14

### LLM
langchain==0.0.197
chromadb==0.3.23
urllib3
PyMuPDF==1.22.3
python-dotenv==1.0.0
unstructured==0.6.6
extract-msg==0.41.1
tabulate==0.9.0
pandoc==2.3
pypandoc==1.11
tqdm==4.65.0
openai
tiktoken
fpdf
